First commit of schema and validator
This is the first attempt at creating a JSON schema for RNAcentral data
import. This schema isn't test yet and may not be complete. It is
currently built off the AGR schemas
(https://github.com/alliance-genome/agr_schemas) as they have already
worked out some of what we need to do. This is hopefully close to a
useful example and schema but I'm not sure yet.

There is also a simple validator which can validate the example. It
likely can be improved as it uses a distinct method of validating if the
secondary structure and sequence have the same length.