Add automated testing
This will now run validation on all example files. The tests are very
minimal right now but this is the start of a more complete testing
strategy we want to take. This uses py.test for testing as it is simple
and useful. We added a configuration file so it finds the expected
files.

In addition, I had to rework the validate script a bit to ensure that it
can be used as a module. The interface is still slightly awkward so
maybe I will work on that more later.