Add validation for name generation
This adds some simple logic for checking if names can be generated.
While RNAcentral requires names for all ncRNA's not all databases
provide them. To deal with this we generate names based off given
information. For example if a ncRNA (or transcript) level name is given
we use that. If one is not present we can fall back to gene level
information. This adds some validation to check that the fall backs
exist if needed. This doesn't truly compute the name, though validation
can log what it would be chosen as a name. As part of this the 'name'
field is no longer required.