Add section for related sequences
This provides a more extensible way of declaring which sequences are
related to which. The idea is to provide a section where any related
sequence can be represented by the sequence id and a relationship. Right
now the only known type of relationship is precursor, but we will add
more in the future. It would be ideal to use an ontology for this, but
for now we will use a enum to create a limited, but controlled
vocabulary.

This approach replaces the previous method of nothing the precursors
of using precursorSequenceId.