Improve the sequence checking constraint
The sequence constraint should both limit the types of characters
allowed as well as limit the fraction of sequences that are uncertain.
This updates the regex to include more allowed characters as well as add
the fraction uncertain constraint. Since this is a new constraint I
added a test to ensure that the validation works.