Update validation script and add new validator
We add a validator that will check if a taxon id is active or obsolete.
This may prove to be too slow on large data sets and in that case we
will modify things to make it optional. Additionally, I rework the
validation script so that it prints out all errors not just the first
that it encounters.