Add evidence to relatedSequences
For TarBase it is important to add information on the evidence that
states two sequence interact. Thus we add an evidence section with a
method field. This should cover their needs and provide us a way to
track this information.

It is currently very simple, no ECO codes but could be extended that
way. Another option would be to use one of the more standard formats
that provide these sorts of annotations if the need arises.