Cleanup genomicLocations
One major issue we have had with genomic locations has been making sure
people know how the coordinates are stored. To deal with this I've added
a field which lets people specify which coordinate system they are
using. For example in recent discussions with PomBase we almost go
coordinates which are off by one. In general catching those mistakes and
fixing this is very time consuming. By letting users specify which of
the two most common ones they are using we can automatically and clearly
get the correct coordinates. Hopefully this is useful to users.

I've also stripped out some unneeded fields from the genomicLocations
section.