Add LncBook example
This represents the basic data we need for importing LncBook. We will
import LncRNAWiki data by using the xref from LncBook. This is somewhat
incomplete in that we do not have all locations of exons which we want.
Nor do we have the disease information that they provide. I will work on
adding that once I read over what AGR does in more detail.