Let mgnify example validate
We are tweaking and making it more complicated to allow mgnify data to
validate. This adds the required changes. The structure of
inferredPhylogeny is changed to a string. This is ok because they are
the only people who use it. Additionally, we add an rfamAccession field,
which should be treated as a source of information such as SO terms.
Finally, I make the schema more complex by allowing it to miss taxonId
if it has inferredPhylogeny and the same for soTermId vs rfamAccession.